requirements.txt
google-adk>=0.1.0
anthropic>=0.25.0
python-dotenv>=1.0.0
jsbeautifier

# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
